requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2

# Spotify API
spotipy==2.23.0

# Google Cloud
google-cloud-firestore==2.13.1
google-cloud-bigquery==3.13.0
google-cloud-storage==2.13.0
google-cloud-aiplatform==1.38.1

# ML & Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
hdbscan==0.8.33
umap-learn==0.5.5

# Audio Processing
librosa==0.10.1
pydub==0.25.1

# API & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0